Renamed file to contexts.py
